Fix crash when loading 4GB+ local files on 32-bit Windows
|total_bytes_to_send| is of type uint64_t, which indicates the size of
the local file (when the Range field is not specified in the headers),
and its value may exceed 4GB. On Windows 32-bit operating systems,
size_t is represented by uint32_t.
`base::checked_cast<size_t>(total_bytes_to_send)` actually checks the
conversion of uint64_t to uint32_t. When the value of
|total_bytes_to_send| exceeds 4GB, it cannot pass the check and will
cause a crash.

Why this change is safe: The maximum data size that
`producer_handle->WriteData` can handle is uint32_t, so the size of
|bytes_to_write| should be guaranteed to be within the range of
uint32_t. We should first take the minimum value of
|bytes_to_write.size()| and |total_bytes_to_send| to ensure that it does
not exceed the size of |bytes_to_write| as the count value of
`bytes_to_write.first`. Finally, the conversion of uint64_t type to
size_t needs to be safety checked, which is reasonable and safe.

Bug: 369739222
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5892980
Auto-Submit: 吴金立 <wujinli@bytedance.com>
Commit-Queue: Takashi Toyoshima <toyoshim@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1366682}

+++ content/browser/loader/file_url_loader_factory.cc
@@ -688,10 +688,11 @@ class FileURLLoader : public network::mojom::URLLoader {
       // applicable. This will always fit in the pipe (see DCHECK above, and
       // assertions near network::features::GetDataPipeDefaultAllocationSize()).
       base::span<const uint8_t> bytes_to_write =
-          base::as_byte_span(initial_read_buffer).subspan(first_byte_to_send);
-      bytes_to_write = bytes_to_write.first(
-          std::min(bytes_to_write.size(),
-                   base::checked_cast<size_t>(total_bytes_to_send)));
+          base::as_byte_span(initial_read_buffer)
+              .subspan(base::checked_cast<size_t>(first_byte_to_send));
+      bytes_to_write = bytes_to_write.first(base::checked_cast<size_t>(std::min(
+          static_cast<uint64_t>(bytes_to_write.size()), total_bytes_to_send)));
+
       size_t actually_written_bytes = 0;
       MojoResult result = producer_handle->WriteData(
           bytes_to_write, MOJO_WRITE_DATA_FLAG_NONE, actually_written_bytes);
